# docs/api/openapi.yaml
components:
  schemas:
...
    ApiMeta:
      additionalProperties: true
      description: >-
        List metadata (docs/07_API.md §19). The server normalizes page/limit (max server-enforced limit: 100) and
        computes total/totalPages/hasNext. Clients must never derive totals themselves.
      properties:
        hasNext:
          type: boolean
        limit:
          maximum: 100
          minimum: 1
          type: integer
        page:
          minimum: 1
          type: integer
        total:
          minimum: 0
          type: integer
        totalPages:
          minimum: 0
          type: integer
      type: object
...
    ConversationDto:
      additionalProperties: false
      description: >-
        The service-request conversation (1:1 with the request). Participants are the request customer and the
        targeted/assigned technician.
      properties:
        createdAt:
          format: date-time
          type: string
        id:
          format: uuid
          type: string
        requestStatus:
          $ref: '#/components/schemas/ServiceRequestStatus'
        serviceRequestId:
          format: uuid
          type: string
      required:
        - id
        - serviceRequestId
        - requestStatus
        - createdAt
      type: object
...
    ErrorCode:
      enum:
        - AUTH_REQUIRED
        - AUTH_INVALID
        - FORBIDDEN
        - NOT_FOUND
        - VALIDATION_ERROR
        - CONFLICT
        - RATE_LIMITED
        - SUBSCRIPTION_REQUIRED
        - ENTITLEMENT_REQUIRED
        - INVALID_STATE_TRANSITION
        - UPLOAD_REJECTED
        - INTERNAL_ERROR
      type: string
    ErrorResponse:
      description: Canonical error envelope (docs/07_API.md §3).
      properties:
        error:
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            fields:
              additionalProperties:
                type: string
              description: Field-level validation details when applicable.
              type: object
            message:
              type: string
          required:
            - code
            - message
          type: object
      required:
        - error
      type: object
...
    MessageDto:
      additionalProperties: false
      description: >-
        Chat message. Ordered newest-first for pagination; the client adapter may reverse for display. readAt exists in
        the model but no mark-read route is documented (deferred).
      properties:
        body:
          type: string
        conversationId:
          format: uuid
          type: string
        createdAt:
          format: date-time
          type: string
        id:
          format: uuid
          type: string
        messageType:
          enum:
            - text
          type: string
        senderUserId:
          format: uuid
          type: string
      required:
        - id
        - conversationId
        - senderUserId
        - messageType
        - body
        - createdAt
      type: object
    NotificationDto:
      additionalProperties: false
      description: "A persisted notification for the authenticated recipient. \type is an open string until product trigger types are ratified; no delivery provider exists (persistence + read APIs only)."
      properties:
        bodyAr:
          type: string
        createdAt:
          format: date-time
          type: string
        dataJson:
          additionalProperties: true
          nullable: true
          type: object
        id:
          format: uuid
          type: string
        readAt:
          format: date-time
          nullable: true
          type: string
        titleAr:
          type: string
        type:
          type: string
      required:
        - id
        - type
        - titleAr
        - bodyAr
        - dataJson
        - readAt
        - createdAt
      type: object
...
    ServiceRequestStatus:
      enum:
        - pending
        - accepted
        - on_the_way
        - in_progress
        - completed
        - cancelled
      type: string
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  description: >-
    Al-Khabir backend contract. Generated from the NestJS controllers (ADR-0003). The canonical success envelope is `{
    data, meta? }`; errors use `{ error: { code, message, fields? } }` with the codes documented in docs/07_API.md §21.
  title: Al-Khabir API
  version: 0.1.0
openapi: 3.0.3
paths:
...
  /api/v1/conversations/{id}/messages:
    get:
      operationId: ChatController_listMessages
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: page
          required: false
          schema:
            minimum: 1
            type: integer
        - in: query
          name: limit
          required: false
          schema:
            maximum: 100
            minimum: 1
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/MessageDto'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                required:
                  - data
                type: object
          description: Bounded message history (newest first).
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation failed (canonical error envelope).
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Missing/malformed token.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conversation not found or not a participant (identical response).
      security:
        - bearer: []
      tags:
        - chat
    post:
      operationId: ChatController_sendMessage
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: false
              properties:
                body:
                  maxLength: 2000
                  minLength: 1
                  type: string
              required:
                - body
              type: object
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/MessageDto'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                required:
                  - data
                type: object
          description: Message created; sender = verified JWT subject.
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation failed (canonical error envelope).
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Missing/malformed token.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conversation not found or not a participant (identical response).
      security:
        - bearer: []
      tags:
        - chat
...
  /api/v1/notifications:
    get:
      operationId: NotificationsController_list
      parameters:
        - in: query
          name: page
          required: false
          schema:
            minimum: 1
            type: integer
        - in: query
          name: limit
          required: false
          schema:
            maximum: 100
            minimum: 1
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/NotificationDto'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                required:
                  - data
                type: object
          description: Own notifications (paginated, newest first).
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation failed (canonical error envelope).
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Missing/malformed token.
      security:
        - bearer: []
      tags:
        - notifications
  /api/v1/notifications/read-all:
    post:
      operationId: NotificationsController_markAllRead
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/ReadAllResultDto'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                required:
                  - data
                type: object
          description: All own unread notifications marked read.
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation failed (canonical error envelope).
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Missing/malformed token.
      security:
        - bearer: []
      tags:
        - notifications
  /api/v1/notifications/{id}/read:
    post:
      operationId: NotificationsController_markRead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/NotificationDto'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                required:
                  - data
                type: object
          description: Notification marked read.
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation failed (canonical error envelope).
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Missing/malformed token.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Notification not found or not owned by this recipient.
      security:
        - bearer: []
      tags:
        - notifications
...
  /api/v1/service-requests/{id}/conversation:
    get:
      operationId: ChatController_getForRequest
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/ConversationDto'
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                required:
                  - data
                type: object
          description: Authorized conversation (lazily created on first access).
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Validation failed (canonical error envelope).
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Missing/malformed token.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conversation not found or not a participant (identical response).
      security:
        - bearer: []
      tags:
        - chat
...